requirements.txt
appdirs==1.4.4
asgiref==3.2.10
backcall==0.2.0
black==20.8b1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
ipdb==0.13.4
ipython==7.16.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
packaging==20.8
parso==0.7.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
Pygments==2.7.3
PyJWT==1.7.1
pyparsing==2.4.7
python-decouple==3.3
pytz==2020.4
regex==2020.11.13
requests==2.25.0
responses==0.12.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
toml==0.10.2
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
validate-docbr==1.8.2
wcwidth==0.2.5
gunicorn==20.0.4
